Fix MockConnectionManager.getConnection() and extract inner classes to its own package
- MockConnectionManager.getConnection() should not create a connection for an already
shutdown instance.
- MockClientConnection.write() should check if server is running
- Refactor TestNodeRegistry inner classes to their own package
