+++ hazelcast-client/src/test/java/com/hazelcast/client/test/TestClientRegistry.java
@@ -41,7 +41,8 @@ import com.hazelcast.nio.ConnectionType;
 import com.hazelcast.nio.OutboundFrame;
 import com.hazelcast.spi.discovery.integration.DiscoveryService;
 import com.hazelcast.spi.impl.NodeEngineImpl;
-import com.hazelcast.test.TestNodeRegistry;
+import com.hazelcast.test.mocknetwork.MockConnection;
+import com.hazelcast.test.mocknetwork.TestNodeRegistry;
 import com.hazelcast.util.ExceptionUtil;
 
 import java.io.IOException;
@@ -169,9 +170,13 @@ public class TestClientRegistry {
 
         @Override
         public boolean write(OutboundFrame frame) {
+            Node node = serverNodeEngine.getNode();
+            if (!node.isRunning()) {
+                return false;
+            }
             ClientMessage newPacket = readFromPacket((ClientMessage) frame);
             lastWriteTime = System.currentTimeMillis();
-            serverNodeEngine.getNode().clientEngine.handleClientMessage(newPacket, serverSideConnection);
+            node.clientEngine.handleClientMessage(newPacket, serverSideConnection);
             return true;
         }
 
@@ -235,7 +240,7 @@ public class TestClientRegistry {
         }
     }
 
-    private class MockedNodeConnection extends TestNodeRegistry.MockConnection {
+    private class MockedNodeConnection extends MockConnection {
 
         private final MockedClientConnection responseConnection;
         private final int connectionId;
